+++ CHANGELOG.md
@@ -12,11 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New example that finetunes a pre-trained ResNet model on the Cats & Dogs dataset.
 - Added a '@requires_gpus' decorator for marking tests as needing GPUs. Tests marked with this will be run in the "GPU Tests" workflow
   on dual k80 GPUs via Beaker.
+- Added the "-w/--workspace" option to `tango run` and `tango server` commands. This option takes a URL, and instantiates the workspace from the URL using the newly added `Workspace.from_url()` method.
+- Added the "workspace" field to `TangoGlobalSettings`.
 - Added a utility function to get a `StepGraph` directly from a file.
 
 ### Changed
 
 - `local_workspace.ExecutorMetadata` renamed to `StepExecutionMetadata` and now saved as `execution-metadata.json`.
+- Renamed `common.logging.file_handler()` to `with_file_handler()`.
+- `tango run` without the option "-w/--workspace" or "-d/--workspace-dir" will now use a `MemoryWorkspace` instead of a `LocalWorkspace` in a temp directory, unless you've specified
+  a default workspace in a `TangoGlobalSettings` file.
 
 ### Fixed